[sigverify] Update SPHINCS+ to use aligned signatures and keys.
Since ROM public keys and signatures are already required to be
word-aligned, we can align a bunch of internal buffers in SPHINCS+ and
remove a cast-to-bytes for sigverify. The performance impact is
basically nil (about 1000 cycles faster out of a total ~1M cycles),
because I think the compiler already noticed that most of these were
aligned.

+++ sw/device/silicon_creator/lib/sigverify/sphincsplus/fors.h
@@ -33,7 +33,7 @@ extern "C" {
  * @return Error code indicating if the operation succeeded.
  */
 OT_WARN_UNUSED_RESULT
-rom_error_t fors_pk_from_sig(const uint8_t *sig, const uint8_t *m,
+rom_error_t fors_pk_from_sig(const uint32_t *sig, const uint8_t *m,
                              const spx_ctx_t *ctx, const spx_addr_t *fors_addr,
                              uint32_t *pk);
