+++ sw/device/silicon_creator/lib/sigverify/sphincsplus/thash.h
@@ -35,7 +35,7 @@ extern "C" {
  * @return Error code indicating if the operation succeeded.
  */
 OT_WARN_UNUSED_RESULT
-rom_error_t thash(const uint8_t *in, size_t inblocks, const spx_ctx_t *ctx,
+rom_error_t thash(const uint32_t *in, size_t inblocks, const spx_ctx_t *ctx,
                   const spx_addr_t *addr, uint32_t *out);
 
 #ifdef __cplusplus
